+++ src/main/java/com/spring/cajero/automatico/model/dtos/DtoMoneda.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class DtoMoneda implements Serializable{
 	
 	private static final long serialVersionUID = 1L;
-	private Integer mo_denominacion;
-	private Integer mo_cantidad;
+	private Long mo_denominacion;
+	private Long mo_cantidad;
 	
 	//--
-	public DtoMoneda(Integer mo_denominacion, Integer mo_cantidad) {
+	public DtoMoneda(Long mo_denominacion, Long mo_cantidad) {
 		super();
 		this.mo_denominacion = mo_denominacion;
 		this.mo_cantidad = mo_cantidad;
@@ -18,17 +18,17 @@ public class DtoMoneda implements Serializable{
 	public DtoMoneda() {}
 	
 	//----
-	public Integer getMo_denominacion() {
+	public Long getMo_denominacion() {
 		return mo_denominacion;
 	}
 
-	public void setMo_denominacion(Integer mo_denominacion) {
+	public void setMo_denominacion(Long mo_denominacion) {
 		this.mo_denominacion = mo_denominacion;
 	}
-	public Integer getMo_cantidad() {
+	public Long getMo_cantidad() {
 		return mo_cantidad;
 	}
-	public void setMo_cantidad(Integer mo_cantidad) {
+	public void setMo_cantidad(Long mo_cantidad) {
 		this.mo_cantidad = mo_cantidad;
 	}
