+++ src/main/java/com/spring/cajero/automatico/model/repositories/IRepositoryEntityManager.java
@@ -10,10 +10,10 @@ public interface IRepositoryEntityManager {
 	
 	
 	public Object listarSaldoDenominacionCantidad();
-	public void updateCantidadbyMonedasBilletes(@PathVariable String tipo, @PathVariable Integer denominacion, @PathVariable Integer cantidad);
-	public  Object retirarSaldo(@PathVariable Integer cantidad);
+	public void updateCantidadbyMonedasBilletes(@PathVariable String tipo, @PathVariable Long denominacion, @PathVariable Long cantidad);
+	public  Object retirarSaldo(@PathVariable Long cantidad);
 	public Object getSaldoActual();
-	public void  updateMonedasBilletes(@PathVariable Integer bi20, @PathVariable Integer bi50 ,@PathVariable Integer bi100, @PathVariable Integer bi200,
-									@PathVariable Integer mo5, @PathVariable Integer mo10, @PathVariable Integer mo20);
+	public void  updateMonedasBilletes(@PathVariable Long bi20, @PathVariable Long bi50 ,@PathVariable Long bi100, @PathVariable Long bi200,
+									@PathVariable Long mo5, @PathVariable Long mo10, @PathVariable Long mo20);
 
 }
