+++ src/main/java/com/spring/cajero/automatico/model/services/IServices.java
@@ -4,9 +4,9 @@ import com.spring.cajero.automatico.model.dtos.DtoRequestUpdateMonedasBilletes;
 
 public interface IServices {
 	
-	public void updateCantidadbyMonedasBilletes(String tipo, Integer denominacion, Integer cantidad);
+	public void updateCantidadbyMonedasBilletes(String tipo, Long denominacion, Long cantidad);
 	public Object listarSaldoDenominacionCantidad();
-	public Object retirarSaldo(Integer cantidad);
+	public Object retirarSaldo(Long cantidad);
 	public Object getSaldoActual();
 	public void updateMonedasBilletes( DtoRequestUpdateMonedasBilletes dto);
